requirements.txt
attrs==24.2.0 ; python_version >= "3.10" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
configargparse==1.7 ; python_version >= "3.10" and python_version < "4.0"
contourpy==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
dash-core-components==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
dash-html-components==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
dash-table==5.0.0 ; python_version >= "3.10" and python_version < "4.0"
dash==2.18.2 ; python_version >= "3.10" and python_version < "4.0"
fast-simplification==0.1.9 ; python_version >= "3.10" and python_version < "4.0"
fastjsonschema==2.20.0 ; python_version >= "3.10" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.54.1 ; python_version >= "3.10" and python_version < "4.0"
idna==3.10 ; python_version >= "3.10" and python_version < "4.0"
imageio==2.36.0 ; python_version >= "3.10" and python_version < "4.0"
importlib-metadata==8.5.0 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.10" and python_version < "4.0"
jupyter-core==5.7.2 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.7 ; python_version >= "3.10" and python_version < "4.0"
lazy-loader==0.4 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.9.2 ; python_version >= "3.10" and python_version < "4.0"
nbformat==5.10.4 ; python_version >= "3.10" and python_version < "4.0"
nest-asyncio==1.6.0 ; python_version >= "3.10" and python_version < "4.0"
networkx==3.4.2 ; python_version >= "3.10" and python_version < "4.0"
numpy==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
open3d==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.10" and python_version < "4.0"
pillow==11.0.0 ; python_version >= "3.10" and python_version < "4.0"
platformdirs==4.3.6 ; python_version >= "3.10" and python_version < "4.0"
plotly==5.24.1 ; python_version >= "3.10" and python_version < "4.0"
pymeshlab==2023.12.post2 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pywin32==308 ; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
retrying==1.3.4 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
scikit-image==0.24.0 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.14.1 ; python_version >= "3.10" and python_version < "4.0"
setuptools==75.4.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
tenacity==9.0.0 ; python_version >= "3.10" and python_version < "4.0"
tifffile==2024.9.20 ; python_version >= "3.10" and python_version < "4.0"
traitlets==5.14.3 ; python_version >= "3.10" and python_version < "4.0"
trimesh==4.5.2 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.10" and python_version < "4.0"
werkzeug==3.0.6 ; python_version >= "3.10" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.10" and python_version < "4.0"
